Resuelto error de la ventana  de muestra
